+++ internal/provider/fwprovider/auth_gcp.go
@@ -19,7 +19,8 @@ func AuthLoginGCPSchema() schema.Block {
 		NestedObject: schema.NestedBlockObject{
 			Attributes: map[string]schema.Attribute{
 				consts.FieldRole: schema.StringAttribute{
-					Required:    true,
+					// can be set via an env var
+					Optional:    true,
 					Description: "Name of the login role.",
 				},
 				consts.FieldJWT: schema.StringAttribute{